requirements.txt
# Core data
pandas
numpy
duckdb
pybaseball

# Visualization
plotly
matplotlib
seaborn
kaleido==0.2.1

# Web apps
streamlit

# Social media
tweepy

# ML (Phase 6)
scikit-learn
scipy

# MLB API
MLB-StatsAPI

# Utilities
requests
python-dotenv

# Notebooks
marimo
